backend/app/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.30.6
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9